wipp advanced events handling

+++ app/api-go/client/client.go
@@ -61,7 +61,7 @@ func WsHandlerCtrl(res http.ResponseWriter, req *http.Request) {
 			log.Println(prefixLog, "command sent;", r.Content, r.Press)
 			player.Mutex.Lock()
 			cmd, _ := strconv.Atoi(r.Content)
-			if int8(cmd) != Keymap.KEY_SPACEBAR {
+			if cmd != Keymap.KEY_SPACEBAR {
 				player.BotContext.Moving = r.Press
 			} else {
 				conn.WriteJSON(&game.TextData{Type: game.TypeAlert, Value: "Robot Firing !"})
